requirements.txt
streamlit
pandas
beautifulsoup4
lxml
requests
google-generativeai
python-dotenv
urllib3